hw/block/nvme: Add support for Namespace Types
Namespace Types introduce a new command set, "I/O Command Sets",
that allows the host to retrieve the command sets associated with
a namespace. Introduce support for the command set and enable
detection for the NVM Command Set.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>
Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>

hw/block/nvme: Add a helper function to zero-fill an identify struct

Zero-filled identify structs are heavily used in the NVMe specification.
E.g., certain CNS commands are defined to return a zero-filled identify
struct when an inactive namespace NSID is supplied.

Add a helper function in order to remove duplicated code.

Signed-off-by: Niklas Cassel <niklas.cassel@wdc.com>
Signed-off-by: Dmitry Fomichev <dmitry.fomichev@wdc.com>